Fixed snapshot loading for DFSA and used-table crawlers (#3046)
## Changes

This PR fixes an issue with the DFSA and used-table crawlers that could
prevent loading of the snapshots. When loading they convert the rows to
dictionaries using `.as_dict()` which isn't available on rows provided
by the spark-based lsql backend. Instead `.asDict()` needs to be used.

Incidental changes:
- An existing integration test was updated to also test snapshot loading
for these crawlers.
 - Another test was renamed to fix a typo in the name.

### Linked issues

Relates to #3036, #3039.

### Tests

- existing unit tests
- existing integration tests