Refactor refreshing of migration-status information for tables, eliminate another redundant refresh. (#3270)
## Changes

This PR updates the way table records are enriched with migration-status
information during encoding for the history log. Changes include:

- Ensuring refresh of the migration-status information is explicit and
under control of the workflow, an intent expressed
[here](https://github.com/databrickslabs/ucx/pull/2743#discussion_r1783326822)
when the current design was laid out.
- A redundant refresh of the migration-status information is eliminated.
(This took place in the `update_tables_history_log` task, even though it
had just occurred in the `refresh_table_migration_status` task
immediately prior to it.)
- Some additional (unit) test coverage of the
`migration-progress-experimental` workflow.

### Linked issues

Splits #3239.

### Functionality

- modified existing workflow: `migration-progress-experimental`

### Tests

- added/updated unit tests
- existing integration tests

---------

Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>