Fix integration test: `test_running_real_assessment_job_ext_hms` (#3421)
## Changes

This PR updates the `test_running_real_assessment_job_ext_hms`
integration test:

- It ensures that the workflow runs on a cluster with an external HMS,
instead of faking it. (The previous method was insufficient once we
switched to the fast table crawler and the Python4J bridge became
needed.)
- Update the test fixture so that the resources it creates to be crawled
are created in the external HMS, not the main one. Without this the
assessment crawlers partially fail because the resources they are
configured to crawl do not exist on the external HMS.

### Linked issues

Resolves #3134.
Resolves #3246.

### Tests

- updated integration test