Release v0.53.1 (#3472)
* Removed `packaging` package dependency
([#3469](https://github.com/databrickslabs/ucx/issues/3469)). In this
release, we have removed the dependency on the `packaging` package in
the open-source library to address a release issue. The import
statements for "packaging.version.Version" and
"packaging.version.InvalidVersion" have been removed. The function
_external_hms in the federation.py file has been updated to retrieve the
Hive Metastore version using the "spark.sql.hive.metastore.version"
configuration key and validate it using a regular expression pattern. If
the version is not valid, the function logs an informational message and
returns None. This change modifies the Hive Metastore version validation
logic and improves the overall reliability and maintainability of the
library.