Update table-migration workflows to also capture updated migration progress into the history log (#3239)
## Changes

The table-migration workflows already contained tasks at the end that
log information about tables that still need to be migrated. The primary
purpose of this PR is to update these workflows so they also capture
updated progress information into the history log.

Other changes include:

 - Updating the documentation for which workflows update which tables.
- ~Updating the (singleton) encoder for table-history so that
initialisation doesn't trigger an implicit refresh of the
`TableMigrationStatus` data. Instead this is controlled at the workflow
level, as intended.~ Moved to #3270.

### Linked issues

~Conflicts with #3200 (will need rebasing).~ (Resolved.)

### Functionality

- updated documentation
- modified existing workflows:

  - `migrate-tables`
  - `migrate-external-hiveserde-tables-in-place-experimental`
  - `migrate-external-tables-ctas`
  - `scan-tables-in-mounts-experimental`
  - `migrate-tables-in-mounts-experimental`

### Tests

- manually tested
- updated and new unit tests
- updated and new integration tests

---------

Co-authored-by: Serge Smertin <259697+nfx@users.noreply.github.com>
Co-authored-by: Cor Zuurmond <jczuurmond@protonmail.com>