Fix issue in installing UCX on UC enabled workspace (#3501)
<!-- REMOVE IRRELEVANT COMMENTS BEFORE CREATING A PULL REQUEST -->
## Changes
<!-- Summary of your changes that are easy to understand. Add
screenshots when necessary -->

This PR updates the UCX policy that is created when installing UCX. It
replaces the policy definition for spark_version from fixed to allowlist
with a default value.
When UC is enabled on a workspaces, the cluster definition take the
value as single_user and user_isolation instead of the
Legacy_Single_User and Legacy_Table_ACL and this is occurring due to the
policy which is overriding this. Changing the sparkversion from fixed to
allowlist solved this issue. This could be a API issue which needs to be
raised separately, but for now it addresses the problem by doing this
simple change

It also updates the job definition to use the default value when not
passed by setting the apply_policy_default_values to true

### Linked issues
<!-- DOC: Link issue with a keyword: close, closes, closed, fix, fixes,
fixed, resolve, resolves, resolved. See
https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword
-->

Resolves #3420 

### Functionality

- [ ] modified existing command: `databricks labs ucx ...`

### Tests
<!-- How is this tested? Please see the checklist below and also
describe any other relevant tests -->

- [ ] updated unit tests
- [ ] updated integration tests
- [ ] static installation test