Release v0.55.0 (#3573)
* Introducing UCX docs!
([#3458](https://github.com/databrickslabs/ucx/pull/3458)). In this
release, we introduced the new documents for UCX, you can find them
here:
[https://databrickslabs.github.io/ucx/](https://databrickslabs.github.io/ucx/)
* Hosted Runner for release
([#3532](https://github.com/databrickslabs/ucx/issues/3532)). In this
release, we have made improvements to the release job's security and
control by moving the release.yml file to a new location within a hosted
runner group labeled "linux-ubuntu-latest." This change ensures that the
release job now runs in a protected runner group, enhancing the overall
security and reliability of the release process. The job's environment
remains set to "release," and it retains the same authentication and
artifact signing permissions as before the move, ensuring a seamless
transition while improving the security and control of the release
process.