Add utilitiy integration test to clean up test users in account (#3665)
## Changes
Add utility integration test to clean up test users in account. The test
users in the account are not deleted by our clean up system. If there
are many users, the API becomes slower, increasing the likelihood of
timeouts for our integration tests. This utility test cleans up the test
users. Given that it might interfere with other tests and its nature to
delete account resource, it is disabled from running by default.

### Linked issues

Resolves #3497
Resolves #3547

### Tests

- [x] added utility integration tests