Rename file language helper function (#3661)
## Changes
Rename and document getting supported file language helper function to
make more clear what it does.

IMO it was not clear that this function contains essential decision
power as it is the filter for passing files to the linter only if they
are supported (in contrast with inferring any language - also the
non-supported ones).

For example, before #3655, the `FileLinter` contains filters for
extensions, however this redundant as that is already handled by this
helper function.

The rename makes it more explicit what this function does. The
documentation makes it more clear how to use the function

### Functionality

- [x] modified linter related code.